#500 Setup aws provider + cleanup plugin configs
