+++ strapi/config/plugins.graphql.config.ts
@@ -1,0 +1,10 @@
+const graphqlConfig = {
+  generateArtifacts: true,
+  artifacts: {
+    // When changing schema path, also change watchIgnoreFiles in strapi/config/admin.js
+    schema: true,
+  },
+  defaultLimit: 100,
+};
+
+export default graphqlConfig;
