+++ strapi/config/plugins.ts
@@ -1,98 +1,25 @@
-/**
- * The indexes that are used in search are stored in one shared index. This wraps them to have a unified way for search
- * and easily unwrappable structure to be used separately.
- */
-const wrapSearchIndexEntry = (type, data, commonAttributes?) => {
-  // Remove when https://github.com/meilisearch/strapi-plugin-meilisearch/pull/554 merged
-  const newData = { ...data };
-  delete newData.createdBy;
-  delete newData.updatedBy;
-
-  return {
-    type,
-    id: data.id, // must be present to work correctly
-    locale: data.locale,
-    // [type] is used instead of "data", to avoid  naming clashes of filterable / sortable / searchable attributes
-    [type]: newData,
-    commonAttributes,
-  };
-};
-
-// Because a bug in Meilisearch shared index, only the last added entity's settings are used and the old ones are overwritten
-// instead of merging. Therefore, for all entities we must provide shared settings.
-// In order to take effect of this settings change you need to unhook and rehook index you are modifying.
-const searchIndexSettings = {
-  searchableAttributes: [
-    // Page
-    "page.title",
-    "page.seo.keywords",
-    "page.perex",
-    // Basic documents
-    "basic-document.title",
-    "basic-document.seo.keywords",
-    "basic-document.description",
-    // Blog post
-    "blog-post.title",
-    "blog-post.seo.keywords",
-    // Documents
-    "document.title",
-    "document.description",
-    // Disclosures
-    "disclosure.title",
-    "disclosure.description",
-    "disclosure.contractor",
-    // Event
-    "event.title",
-    "event.seo.keywords",
-    "event.description",
-    // Notice
-    "notice.title",
-    "notice.seo.keywords",
-    "notice.body",
-    // Branch
-    "branch.title",
-    "branch.seo.keywords",
-    "branch.body",
-  ],
-  filterableAttributes: [
-    // All
-    "type",
-    // Page + Event`
-    "locale",
-    // Basic document
-    "basic-document.file_category.id",
-    // Document
-    "document.documentCategory.id",
-    // Disclosure
-    "disclosure.type",
-    // Event
-    "event.dateFromTimestamp",
-    "event.dateToTimestamp",
-    "event.eventTagsIds",
-    "event.eventCategory.id",
-    "event.branch.id",
-    "event.locale",
-  ],
-  sortableAttributes: [
-    // Basic document
-    "basic-document.date_added",
-    // Event
-    "event.dateFromTimestamp",
-    // Document, Disclosure
-    "commonAttributes.addedAtTimestamp",
-  ],
-  pagination: {
-    // https://docs.meilisearch.com/learn/advanced/known_limitations.html#maximum-number-of-results-per-search
-    maxTotalHits: 100000,
-  },
-};
+import meilisearchConfig from "./plugins.meilisearch.config";
+import graphqlConfig from "./plugins.graphql.config";
 
 export default ({ env }) => ({
   graphql: {
+    config: graphqlConfig,
+  },
+  meilisearch: {
+    config: meilisearchConfig,
+  },
+  email: {
     config: {
-      defaultLimit: 100,
-      artifacts: {
-        schema: true,
+      //  mailgun email provider https://www.npmjs.com/package/@strapi/provider-email-mailgun
+      provider: "mailgun",
+      providerOptions: {
+        key: env("MAILGUN_API_KEY"), // Required
+        domain: env("MAILGUN_DOMAIN"), // Required
+        url: "https://api.eu.mailgun.net", //Optional. If domain region is Europe use 'https://api.eu.mailgun.net'
+      },
+      settings: {
+        defaultFrom: env("MAILGUN_EMAIL"),
+        defaultReplyTo: env("MAILGUN_EMAIL"),
       },
     },
   },
@@ -116,122 +43,6 @@ export default ({ env }) => ({
           pathAttribute: "newSlug",
         },
       ],
-    },
-  },
-  email: {
-    config: {
-      //  mailgun email provider https://www.npmjs.com/package/@strapi/provider-email-mailgun
-      provider: "mailgun",
-      providerOptions: {
-        key: env("MAILGUN_API_KEY"), // Required
-        domain: env("MAILGUN_DOMAIN"), // Required
-        url: "https://api.eu.mailgun.net", //Optional. If domain region is Europe use 'https://api.eu.mailgun.net'
-      },
-      settings: {
-        defaultFrom: env("MAILGUN_EMAIL"),
-        defaultReplyTo: env("MAILGUN_EMAIL"),
-      },
-    },
-  },
-  meilisearch: {
-    config: {
-      host: process.env.MEILISEARCH_HOST,
-      apiKey: process.env.MEILISEARCH_ADMIN_API_KEY,
-
-      page: {
-        indexName: "search_index",
-        entriesQuery: {
-          locale: "all",
-        },
-        settings: searchIndexSettings,
-        transformEntry: ({ entry }) => wrapSearchIndexEntry("page", entry),
-      },
-
-      "blog-post": {
-        indexName: "search_index",
-        settings: searchIndexSettings,
-        transformEntry: ({ entry }) => wrapSearchIndexEntry("blog-post", entry),
-      },
-
-      document: {
-        indexName: "search_index",
-        settings: searchIndexSettings,
-        transformEntry: ({ entry }) =>
-          wrapSearchIndexEntry(
-            "document",
-            {
-              ...entry,
-            },
-            {
-              // Meilisearch doesn't support filtering dates as ISO strings, therefore we convert it to UNIX timestamp to
-              // use (number) filters.
-              // Transforming publishedAt to addedAtTimestamp to be able to use the same sort as for Disclosures.
-              addedAtTimestamp: entry.publishedAt
-                ? new Date(entry.publishedAt).getTime()
-                : undefined,
-            },
-          ),
-      },
-
-      disclosure: {
-        indexName: "search_index",
-        settings: searchIndexSettings,
-        transformEntry: ({ entry }) =>
-          wrapSearchIndexEntry(
-            "disclosure",
-            {
-              ...entry,
-            },
-            {
-              // Meilisearch doesn't support filtering dates as ISO strings, therefore we convert it to UNIX timestamp to
-              // use (number) filters.
-              addedAtTimestamp: entry.addedAt
-                ? new Date(entry.addedAt).getTime()
-                : undefined,
-            },
-          ),
-      },
-
-      event: {
-        indexName: "search_index",
-        entriesQuery: {
-          locale: "all",
-        },
-        settings: searchIndexSettings,
-        transformEntry: ({ entry }) =>
-          wrapSearchIndexEntry("event", {
-            ...entry,
-            // Meilisearch doesn't support filtering dates as ISO strings, therefore we convert it to UNIX timestamp to
-            // use (number) filters.
-            dateFromTimestamp: entry.dateFrom
-              ? new Date(entry.dateFrom).getTime()
-              : undefined,
-            dateToTimestamp: entry.dateTo
-              ? new Date(entry.dateTo).getTime()
-              : undefined,
-            // It is not possible to filter nested object in arrays in Meilisearch, so we map it to a basic array with
-            // string values.
-            eventTagsIds: (entry.eventTags ?? []).map(({ id }) => id),
-          }),
-      },
-
-      notice: {
-        indexName: "search_index",
-        entriesQuery: {
-          locale: "all",
-        },
-        settings: searchIndexSettings,
-        transformEntry: ({ entry }) => wrapSearchIndexEntry("notice", entry),
-      },
-
-      branch: {
-        indexName: "search_index",
-        entriesQuery: {
-          locale: "all",
-        },
-        settings: searchIndexSettings,
-        transformEntry: ({ entry }) => wrapSearchIndexEntry("branch", entry),
-      },
     },
   },
 });
