+++ strapi/kubernetes/base/.env
@@ -1,8 +1,13 @@
 NODE_OPTIONS="--max-old-space-size=7000"
 PORT=1337
-MINIO_BUCKET=strapi-city-library
+
+MINIO_BUCKET=TODO
+MINIO_PRIVATE_ENDPOINT=https://s3.bratislava.sk
+MINIO_PUBLIC_ENDPOINT=https://TODO.s3.bratislava.sk
+
 STRAPI_PLUGIN_I18N_INIT_LOCALE_CODE=sk
 
 MAILGUN_DOMAIN=no-reply-strapi.bratislava.sk
 MAILGUN_EMAIL=info@no-reply-strapi.bratislava.sk
+
 REVALIDATE_NEXT_URL=http://city-library-next-app
