+++ strapi/package.json
@@ -17,6 +17,7 @@
     "@strapi/plugin-i18n": "4.25.22",
     "@strapi/plugin-users-permissions": "4.25.22",
     "@strapi/strapi": "4.25.22",
+    "@strapi/provider-upload-aws-s3": "4.25.22",
     "@strapi/provider-email-mailgun": "4.25.22",
     "pg": "^8.13.1",
     "postinstall-postinstall": "^2.1.0",
@@ -24,7 +25,6 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "5.3.4",
     "strapi-plugin-meilisearch": "^0.9.2",
-    "strapi-provider-upload-ts-minio": "^3.1.0",
     "styled-components": "^5.2.1"
   },
   "author": {
